feat: update test and add simplecov
