+++ .simplecov
@@ -1,0 +1,1 @@
+SimpleCov.start 'rails'
