+++ spec/controllers/assistants/assistants_controller_show_get_spec.rb
