+++ spec/controllers/assistants/assistants_controller_update_spec.rb
@@ -18,5 +18,12 @@ RSpec.describe AssistantsController, type: :controller do
         put :update, params: { id: "invalid_id", assistant: { instructions: "You are chicken" } }
       end.to raise_error(ActiveRecord::RecordNotFound)
     end
+
+    it "update db failed" do
+      assistant = create(:assistant, member: member)
+      allow_any_instance_of(Assistant).to receive(:update).and_return(false)
+      put :update, params: { id: assistant.id, assistant: { instructions: "You are chicken" } }
+      expect(response).to have_http_status(:unprocessable_entity)
+    end
   end
 end
